+++ src/modules/jwt-manager/jwt-manager.service.ts
@@ -1,17 +1,27 @@
 import { Injectable } from '@nestjs/common';
-import { UserAccessTokenPayload, UserRefreshTokenPayload } from '../auth/auth.interface';
+import {
+    UserAccessTokenPayload,
+    UserAccessTokenPayloadDto,
+    UserRefreshTokenPayload,
+    UserRefreshTokenPayloadDto
+} from './jwt-manager.types';
 import { JwtService } from '@nestjs/jwt';
+import { generateUserDataPayload } from './jwt-manager.utils';
 
 @Injectable()
 export class JwtManagerService {
 
     constructor(private readonly jwtService: JwtService) {}
 
-    public async generateAccessToken(userDataPayload: UserAccessTokenPayload): Promise<string> {
+    public async generateAccessToken(userDataPayloadDto: UserAccessTokenPayloadDto): Promise<string> {
+        const userDataPayload = generateUserDataPayload(userDataPayloadDto);
+
         return await this.jwtService.signAsync(userDataPayload, { secret: process.env.ACCESS_TOKEN_SECRET });
     }
 
-    public async generateRefreshToken(userDataPayload: UserRefreshTokenPayload): Promise<string> {
+    public async generateRefreshToken(userDataPayloadDto: UserRefreshTokenPayloadDto): Promise<string> {
+        const userDataPayload = generateUserDataPayload(userDataPayloadDto);
+
         return await this.jwtService.signAsync(userDataPayload, { secret: process.env.REFRESH_TOKEN_SECRET });
     }
 
@@ -23,10 +33,12 @@ export class JwtManagerService {
         return await this.jwtService.verifyAsync(refreshToken, { secret: process.env.REFRESH_TOKEN_SECRET });
     }
 
+    // TODO: Deprecated. Remove it
     public async encodeUserData(userDataPayload: UserAccessTokenPayload): Promise<string> {
         return await this.jwtService.signAsync(userDataPayload, { secret: process.env.ACCESS_TOKEN_SECRET });
     }
 
+    // TODO: Deprecated. Remove it
     public decodeUserData(jwtCookie: string): UserAccessTokenPayload {
         return this.jwtService.decode(jwtCookie) as UserAccessTokenPayload;
     }