+++ test/user.e2e-spec.ts
@@ -9,7 +9,6 @@ import { LoggerService } from '../src/modules/logger/logger.service';
 import { MailManagerService } from '../src/modules/mail-manager/mail-manager.service';
 import { UserRepository } from '../src/mongodb/repositories/user.repository';
 import { RedisService } from '../src/modules/redis/redis.service';
-import { AxiosService } from '../src/services/axios.service';
 
 describe('UserController (e2e)', () => {
     let app: INestApplication;
@@ -17,7 +16,6 @@ describe('UserController (e2e)', () => {
     let userRepository: UserRepository;
     let jwtManagerService: JwtManagerService;
     let redisService: RedisService;
-    let axiosService: AxiosService;
 
     const getCookie = (res, cookieName) => {
         const cookies = {};
@@ -52,11 +50,9 @@ describe('UserController (e2e)', () => {
         set: jest.fn().mockImplementation((x, y, z, v) => {}),
         get: jest.fn().mockImplementation((x) => x),
         setTokens: jest.fn(),
-        unsetTokens: jest.fn()
-    };
-
-    const mockAxiosServiceProvider = {
-        get: jest.fn()
+        unsetTokens: jest.fn(),
+        getAccessToken: jest.fn(),
+        getRefreshToken: jest.fn()
     };
 
     beforeEach(async () => {
@@ -68,7 +64,6 @@ describe('UserController (e2e)', () => {
             .overrideProvider(MailManagerService).useValue({ sendActivationMail: jest.fn((email, id) => {}) })
             .overrideProvider(JwtManagerService).useValue(mockJwtServiceProvider)
             .overrideProvider(RedisService).useValue(mockRedisServiceProvider)
-            .overrideProvider(AxiosService).useValue(mockAxiosServiceProvider)
             .compile();
 
         app = moduleRef.createNestApplication();
@@ -79,7 +74,6 @@ describe('UserController (e2e)', () => {
         userRepository = moduleRef.get(UserRepository);
         jwtManagerService = moduleRef.get(JwtManagerService);
         redisService = moduleRef.get(RedisService);
-        axiosService = moduleRef.get(AxiosService);
     });
 
     it('/users/create (POST)', () => {
@@ -149,6 +143,11 @@ describe('UserController (e2e)', () => {
                 expect(accessCookie).toBeUndefined();
                 expect(refreshCookie).toBeUndefined();
             });
+    });
+
+    // TODO: It used to work too long, therefore test will be designed in the future
+    describe('/users/refreshTokens (POST)', () => {
+        it('should generate new accessToken', () => {});
     });
 
     describe('/users/:login/grant/:capability (POST)', () => {